+++ indico/modules/events/fields.py
@@ -108,10 +108,14 @@ class PersonLinkListFieldBase(PrincipalListField):
         return persons_settings.get(self.event, 'default_search_external')
 
     @property
-    def can_enter_manually(self):
+    def enforce_user_search(self):
         if self.event is None:
-            return True
-        return self.event.can_manage(session.user) or not persons_settings.get(self.event, 'disallow_custom_persons')
+            return False
+        return persons_settings.get(self.event, 'enforce_user_search')
+
+    @property
+    def required_person_fields(self):
+        return values_from_signal(signals.event.person_required_fields.send(self.get_form()), multi_value_types=list)
 
     @property
     def name_format(self):
@@ -136,14 +140,18 @@ class PersonLinkListFieldBase(PrincipalListField):
         identifier = data.get('identifier')
         affiliations_disabled = self.extra_params.get('disable_affiliations', False)
         data = PersonLinkSchema(unknown=EXCLUDE).load(data)
-        if not self.can_enter_manually and not data.get('type'):
-            raise UserValueError('Manually entered persons are not allowed')
+        if not self.enforce_user_search and not data.get('type'):
+            required_fields = values_from_signal(signals.event.person_required_fields.send(self.get_form()),
+                                                 multi_value_types=list)
+            if any(not data.get(field) for field in required_fields):
+                print(required_fields, data, [field for field in required_fields if not data.get(field)])
+                raise UserValueError('Missing required person fields')
         if identifier and identifier.startswith('ExternalUser:'):
             # if the data came from an external user, look up their affiliation if the names still match;
             # we do not have an affiliation ID yet since it may not exist in the local DB yet
             cache = make_scoped_cache('external-user')
             external_user_data = cache.get(identifier.removeprefix('ExternalUser:'), {})
-            if not self.can_enter_manually:
+            if not self.enforce_user_search:
                 for key in ('first_name', 'last_name', 'email', 'affiliation', 'phone', 'address'):
                     data[key] = external_user_data.get(key, '')
                 data['_title'] = UserTitle.none
@@ -164,7 +172,7 @@ class PersonLinkListFieldBase(PrincipalListField):
             person_link = self.person_link_cls.query.filter_by(person=person, object=self.object).first()
         if not person_link:
             person_link = self.person_link_cls(person=person)
-        if not self.can_enter_manually:
+        if not self.enforce_user_search:
             person_link.populate_from_dict(data, keys=('display_order',))
             return person_link
         person_link.populate_from_dict(data, keys=('first_name', 'last_name', 'affiliation', 'affiliation_link',